Create README.md
Greetings, I'm Arsh Verma, a first year undergrad at SRMIST, and I'm dedicated to making waves in the professional world.
My journey, which began in small town in Gaziabad and led me to Chennai for my UG, is driven by an unwavering commitment to drive meaningful change. My relentless drive gave rise to our startup, a testament to my dedication to positive impact.

In addition to my entrepreneurial pursuits, I have been connected with many upcoming and ongoing startups and have shared a chance to work with my fellow collogues in this entrepreneurial journey.

Follow me on LinkedIn to stay updated on my journey and insights into the world of education, entrepreneurship, and meaningful change. Connect with me today for a future filled with inspiration and innovation!